Multiplayer folder restructuring.
