+++ osu.Game.Tests/Visual/TestCaseMultiHeader.cs
@@ -5,6 +5,7 @@ using NUnit.Framework;
 using osu.Framework.Graphics;
 using osu.Game.Graphics.UserInterface;
 using osu.Game.Screens.Multiplayer;
+using osu.Game.Screens.Multiplayer.Screens;
 
 namespace osu.Game.Tests.Visual
 {
