+++ osu.Game/Screens/Multiplayer/Header.cs
@@ -12,6 +12,7 @@ using osu.Game.Graphics;
 using osu.Game.Graphics.Sprites;
 using osu.Game.Graphics.UserInterface;
 using osu.Game.Overlays.SearchableList;
+using osu.Game.Screens.Multiplayer.Screens;
 using OpenTK;
 using OpenTK.Graphics;
