+++ osu.Game/Screens/Multiplayer/Multi.cs
@@ -8,6 +8,7 @@ using osu.Framework.Screens;
 using osu.Game.Graphics;
 using osu.Game.Graphics.Backgrounds;
 using osu.Game.Graphics.Containers;
+using osu.Game.Screens.Multiplayer.Screens.Lounge;
 
 namespace osu.Game.Screens.Multiplayer
 {
